// src/apis-client/models/StorageRateInput.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Aeler demo
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    DailyRate,
    DailyRateFromJSON,
    DailyRateFromJSONTyped,
    DailyRateToJSON,
    StorageRateBase,
    StorageRateBaseFromJSON,
    StorageRateBaseFromJSONTyped,
    StorageRateBaseToJSON,
} from './';

/**
 * 
 * @export
 * @interface StorageRateInput
 */
export interface StorageRateInput {
    /**
     * 
     * @type {StorageRateBase}
     * @memberof StorageRateInput
     */
    input: StorageRateBase;
    /**
     * 
     * @type {DailyRate}
     * @memberof StorageRateInput
     */
    dailyRate: DailyRate;
}

export function StorageRateInputFromJSON(json: any): StorageRateInput {
    return StorageRateInputFromJSONTyped(json, false);
}

export function StorageRateInputFromJSONTyped(json: any, ignoreDiscriminator: boolean): StorageRateInput {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'input': StorageRateBaseFromJSON(json['input']),
        'dailyRate': DailyRateFromJSON(json['dailyRate']),
    };
}

export function StorageRateInputToJSON(value?: StorageRateInput | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'input': StorageRateBaseToJSON(value.input),
        'dailyRate': DailyRateToJSON(value.dailyRate),
    };
}